Move validation to validate
