+++ pkg/validate/validate_name.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package validation
+package validate
 
 import (
 	"k8s.io/apimachinery/pkg/util/validation"
